Linear Discriminant Analysis using the Generalized Singular Value Decomposition

Fit an LDA/GSVD model.

Usage

ldaGSVD(formula, data)

Arguments

formula

an object of class formula, which has the form class ~ x1 + x2 + ...

data

a data frame that contains both predictors and the response. Missing values are NOT allowed.

Value

An object of class ldaGSVD containing the following components:

scaling: a matrix which transforms the training data to LD scores, normalized so that the within-group scatter matrix is proportional to the identity matrix.

formula: the formula passed to the ldaGSVD()

terms: a object of class terms derived using the input formula and the training data

prior: a table of the estimated prior probabilities.

groupMeans: a matrix that records the group means of the training data on the transformed LD scores.

xlevels: a list records the levels of the factor predictors, derived using the input formula and the training data

Details

Traditional Fisher's Linear Discriminant Analysis (LDA) ceases to work when the within-class scatter matrix is singular. The Generalized Singular Value Decomposition (GSVD) is used to address this issue. GSVD simultaneously diagonalizes both the within-class and between-class scatter matrices without the need to invert a singular matrix. This method is believed to be more accurate than PCA-LDA (as in MASS::lda) because it also considers the information in the between-class scatter matrix.

References

Ye, J., Janardan, R., Park, C. H., & Park, H. (2004). An optimization criterion for generalized discriminant analysis on undersampled problems. IEEE Transactions on Pattern Analysis and Machine Intelligence

Howland, P., Jeon, M., & Park, H. (2003). Structure preserving dimension reduction for clustered text data based on the generalized singular value decomposition. SIAM Journal on Matrix Analysis and Applications

Examples

fit <- ldaGSVD(Species~., data = iris)
# prediction
predict(fit,iris)
#>   [1] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>   [6] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [11] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [16] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [21] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [26] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [31] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [36] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [41] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [46] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [51] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [56] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [61] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [66] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [71] "virginica"  "versicolor" "versicolor" "versicolor" "versicolor"
#>  [76] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [81] "versicolor" "versicolor" "versicolor" "virginica"  "versicolor"
#>  [86] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [91] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [96] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#> [101] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [106] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [111] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [116] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [121] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [126] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [131] "virginica"  "virginica"  "virginica"  "versicolor" "virginica" 
#> [136] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [141] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [146] "virginica"  "virginica"  "virginica"  "virginica"  "virginica"