Predictions from a fitted ldaGSVD object

Prediction of test data using a fitted ldaGSVD object

Usage

# S3 method for ldaGSVD
predict(object, newdata, type = c("response", "prob"), ...)

Arguments

object

a fitted model object of class ldaGSVD, which is assumed to be the result of the ldaGSVD() function.

newdata

data frame containing the values at which predictions are required. Missing values are NOT allowed.

type

character string denoting the type of predicted value returned. The default is to return the predicted class (type = 'response'). The predicted posterior probabilities for each class will be returned if type = 'prob'.

...

further arguments passed to or from other methods.

Value

The function returns different values based on the type, if

type = 'response': vector of predicted responses.

type = 'prob': a data frame of the posterior probabilities. Each class takes a column.

Details

Unlike the original paper, which uses the k-nearest neighbor (k-NN) as the classifier, we use a faster and more straightforward likelihood-based method. One limitation of the traditional likelihood-based method for LDA is that it ceases to work when there are Linear Discriminant (LD) directions with zero variance in the within-class scatter matrix. However, when using LDA/GSVD, all chosen LD directions possess non-zero variance in the between-class scatter matrix. This implies that LD directions with zero variance in the within-class scatter matrix will yield the highest Fisher's ratio. Therefore, to get these directions higher weights, we manually adjust the zero variance to 1e-15 for computational reasons.

References

Ye, J., Janardan, R., Park, C. H., & Park, H. (2004). An optimization criterion for generalized discriminant analysis on undersampled problems. IEEE Transactions on Pattern Analysis and Machine Intelligence

Howland, P., Jeon, M., & Park, H. (2003). Structure preserving dimension reduction for clustered text data based on the generalized singular value decomposition. SIAM Journal on Matrix Analysis and Applications

Examples

fit <- ldaGSVD(Species~., data = iris)
predict(fit,iris)
#>   [1] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>   [6] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [11] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [16] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [21] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [26] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [31] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [36] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [41] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [46] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [51] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [56] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [61] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [66] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [71] "virginica"  "versicolor" "versicolor" "versicolor" "versicolor"
#>  [76] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [81] "versicolor" "versicolor" "versicolor" "virginica"  "versicolor"
#>  [86] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [91] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [96] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#> [101] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [106] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [111] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [116] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [121] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [126] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [131] "virginica"  "virginica"  "virginica"  "versicolor" "virginica" 
#> [136] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [141] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [146] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
# output prosterior probabilities
predict(fit,iris,type = "prob")
#>           setosa   versicolor    virginica
#> 1   1.000000e+00 3.896358e-22 2.611168e-42
#> 2   1.000000e+00 7.217970e-18 5.042143e-37
#> 3   1.000000e+00 1.463849e-19 4.675932e-39
#> 4   1.000000e+00 1.268536e-16 3.566610e-35
#> 5   1.000000e+00 1.637387e-22 1.082605e-42
#> 6   1.000000e+00 3.883282e-21 4.566540e-40
#> 7   1.000000e+00 1.113469e-18 2.302608e-37
#> 8   1.000000e+00 3.877586e-20 1.074496e-39
#> 9   1.000000e+00 1.902813e-15 9.482936e-34
#> 10  1.000000e+00 1.111803e-18 2.724060e-38
#> 11  1.000000e+00 1.185277e-23 3.237084e-44
#> 12  1.000000e+00 1.621649e-18 1.833201e-37
#> 13  1.000000e+00 1.459225e-18 3.262506e-38
#> 14  1.000000e+00 1.117219e-19 1.316642e-39
#> 15  1.000000e+00 5.487399e-30 1.531265e-52
#> 16  1.000000e+00 1.261505e-27 2.268705e-48
#> 17  1.000000e+00 6.754338e-25 3.868271e-45
#> 18  1.000000e+00 4.223741e-21 1.224313e-40
#> 19  1.000000e+00 1.774911e-22 2.552153e-42
#> 20  1.000000e+00 2.593237e-22 5.792079e-42
#> 21  1.000000e+00 1.274639e-19 4.357774e-39
#> 22  1.000000e+00 1.465999e-20 1.987241e-39
#> 23  1.000000e+00 6.569280e-25 7.769177e-46
#> 24  1.000000e+00 8.912348e-15 9.178624e-32
#> 25  1.000000e+00 1.070702e-15 1.167516e-33
#> 26  1.000000e+00 2.497339e-16 5.710269e-35
#> 27  1.000000e+00 3.967732e-17 4.378624e-35
#> 28  1.000000e+00 1.548165e-21 1.595360e-41
#> 29  1.000000e+00 9.271847e-22 6.297955e-42
#> 30  1.000000e+00 9.665144e-17 2.977974e-35
#> 31  1.000000e+00 2.299936e-16 7.182666e-35
#> 32  1.000000e+00 1.975404e-19 2.788334e-38
#> 33  1.000000e+00 7.100041e-27 2.216408e-48
#> 34  1.000000e+00 1.610295e-28 2.743783e-50
#> 35  1.000000e+00 1.205219e-17 1.277245e-36
#> 36  1.000000e+00 1.597186e-21 9.033772e-42
#> 37  1.000000e+00 1.939869e-24 1.662808e-45
#> 38  1.000000e+00 3.310234e-23 7.004971e-44
#> 39  1.000000e+00 4.190242e-17 6.991441e-36
#> 40  1.000000e+00 1.769359e-20 3.541694e-40
#> 41  1.000000e+00 1.063014e-21 2.003866e-41
#> 42  1.000000e+00 2.174217e-11 1.213781e-28
#> 43  1.000000e+00 1.540753e-18 1.305719e-37
#> 44  1.000000e+00 8.940589e-16 1.315511e-32
#> 45  1.000000e+00 1.616206e-17 3.205992e-35
#> 46  1.000000e+00 1.714743e-16 7.172435e-35
#> 47  1.000000e+00 2.083089e-22 2.289783e-42
#> 48  1.000000e+00 2.793482e-18 2.629539e-37
#> 49  1.000000e+00 2.597560e-23 9.820820e-44
#> 50  1.000000e+00 2.322258e-20 4.241757e-40
#> 51  1.969732e-18 9.998894e-01 1.105878e-04
#> 52  1.242878e-19 9.992575e-01 7.425297e-04
#> 53  2.088263e-22 9.958069e-01 4.193053e-03
#> 54  2.198898e-22 9.996423e-01 3.576502e-04
#> 55  4.213678e-23 9.955903e-01 4.409655e-03
#> 56  8.127287e-23 9.985020e-01 1.497982e-03
#> 57  3.549900e-22 9.858346e-01 1.416542e-02
#> 58  5.007065e-14 9.999999e-01 1.119811e-07
#> 59  5.683334e-20 9.998781e-01 1.218649e-04
#> 60  1.241039e-20 9.995027e-01 4.973085e-04
#> 61  1.956628e-18 9.999986e-01 1.420841e-06
#> 62  5.968900e-20 9.992294e-01 7.705716e-04
#> 63  2.716128e-18 9.999988e-01 1.220169e-06
#> 64  1.184445e-23 9.943267e-01 5.673286e-03
#> 65  5.574931e-14 9.999984e-01 1.649215e-06
#> 66  2.369511e-17 9.999573e-01 4.268212e-05
#> 67  8.429328e-24 9.806471e-01 1.935289e-02
#> 68  2.505072e-16 9.999991e-01 9.151716e-07
#> 69  1.670352e-27 9.595735e-01 4.042653e-02
#> 70  1.341503e-17 9.999967e-01 3.296105e-06
#> 71  7.408118e-28 2.532282e-01 7.467718e-01
#> 72  9.399292e-17 9.999907e-01 9.345291e-06
#> 73  7.674672e-29 8.155328e-01 1.844672e-01
#> 74  2.683018e-22 9.995723e-01 4.277469e-04
#> 75  7.813875e-18 9.999758e-01 2.421458e-05
#> 76  2.073207e-18 9.999171e-01 8.290530e-05
#> 77  6.357538e-23 9.982541e-01 1.745936e-03
#> 78  5.639473e-27 6.892131e-01 3.107869e-01
#> 79  3.773528e-23 9.925169e-01 7.483138e-03
#> 80  9.555338e-12 1.000000e+00 1.910624e-08
#> 81  1.022109e-17 9.999970e-01 3.007748e-06
#> 82  9.648075e-16 9.999997e-01 3.266704e-07
#> 83  1.616405e-16 9.999962e-01 3.778441e-06
#> 84  4.241952e-32 1.433919e-01 8.566081e-01
#> 85  1.724514e-24 9.635576e-01 3.644242e-02
#> 86  1.344746e-20 9.940401e-01 5.959931e-03
#> 87  3.304868e-21 9.982223e-01 1.777672e-03
#> 88  2.034571e-23 9.994557e-01 5.443096e-04
#> 89  5.806986e-18 9.999486e-01 5.137101e-05
#> 90  5.981190e-21 9.998183e-01 1.816870e-04
#> 91  5.878614e-23 9.993856e-01 6.144200e-04
#> 92  5.399006e-22 9.980934e-01 1.906591e-03
#> 93  3.559507e-18 9.999887e-01 1.128570e-05
#> 94  2.104146e-14 9.999999e-01 1.135016e-07
#> 95  4.700877e-21 9.996980e-01 3.020226e-04
#> 96  1.584328e-17 9.999817e-01 1.826327e-05
#> 97  2.802293e-19 9.998892e-01 1.108315e-04
#> 98  1.626918e-18 9.999536e-01 4.640488e-05
#> 99  7.638378e-11 1.000000e+00 1.867332e-08
#> 100 4.679301e-19 9.999269e-01 7.305863e-05
#> 101 7.503075e-52 7.127303e-09 1.000000e+00
#> 102 5.213802e-38 1.078251e-03 9.989217e-01
#> 103 1.231264e-42 2.592826e-05 9.999741e-01
#> 104 1.537499e-38 1.068139e-03 9.989319e-01
#> 105 6.242501e-46 1.812963e-06 9.999982e-01
#> 106 4.209281e-49 6.656263e-07 9.999993e-01
#> 107 3.797837e-33 4.862025e-02 9.513797e-01
#> 108 1.352176e-42 1.395463e-04 9.998605e-01
#> 109 1.323390e-42 2.235313e-04 9.997765e-01
#> 110 3.453358e-46 1.727277e-07 9.999998e-01
#> 111 5.452660e-32 1.305353e-02 9.869465e-01
#> 112 1.182560e-37 1.673875e-03 9.983261e-01
#> 113 5.204321e-39 2.006352e-04 9.997994e-01
#> 114 1.269953e-40 1.948672e-04 9.998051e-01
#> 115 1.685361e-45 1.000455e-06 9.999990e-01
#> 116 5.141640e-40 2.605493e-05 9.999739e-01
#> 117 1.909820e-35 6.083553e-03 9.939164e-01
#> 118 1.207799e-44 1.503799e-06 9.999985e-01
#> 119 3.181265e-59 1.317279e-09 1.000000e+00
#> 120 1.598511e-33 2.207990e-01 7.792010e-01
#> 121 1.119461e-42 6.451865e-06 9.999935e-01
#> 122 3.038170e-37 8.272676e-04 9.991727e-01
#> 123 6.032879e-50 9.509838e-07 9.999990e-01
#> 124 1.951261e-31 9.711942e-02 9.028806e-01
#> 125 1.956408e-39 8.836845e-05 9.999116e-01
#> 126 1.109337e-36 2.679310e-03 9.973207e-01
#> 127 7.841997e-30 1.883675e-01 8.116325e-01
#> 128 7.964690e-30 1.342431e-01 8.657569e-01
#> 129 6.190641e-44 1.303681e-05 9.999870e-01
#> 130 1.406448e-32 1.036823e-01 8.963177e-01
#> 131 4.108129e-42 1.442338e-04 9.998558e-01
#> 132 1.555697e-36 5.198047e-04 9.994802e-01
#> 133 1.320330e-45 3.014091e-06 9.999970e-01
#> 134 1.283891e-28 7.293881e-01 2.706119e-01
#> 135 1.926560e-35 6.602253e-02 9.339775e-01
#> 136 1.271083e-45 2.152818e-06 9.999978e-01
#> 137 3.038963e-44 8.881859e-07 9.999991e-01
#> 138 4.605973e-35 6.165648e-03 9.938344e-01
#> 139 4.538634e-29 1.925262e-01 8.074738e-01
#> 140 2.140232e-36 8.290895e-04 9.991709e-01
#> 141 6.570902e-45 1.180810e-06 9.999988e-01
#> 142 6.202588e-36 4.276398e-04 9.995724e-01
#> 143 5.213802e-38 1.078251e-03 9.989217e-01
#> 144 1.073945e-45 1.028519e-06 9.999990e-01
#> 145 4.048249e-46 2.524984e-07 9.999997e-01
#> 146 4.970070e-39 7.473361e-05 9.999253e-01
#> 147 4.616611e-36 5.898784e-03 9.941012e-01
#> 148 5.548962e-35 3.145874e-03 9.968541e-01
#> 149 1.613687e-40 1.257468e-05 9.999874e-01
#> 150 2.858012e-33 1.754229e-02 9.824577e-01