Predictions from a fitted Treee object

Prediction of test data using a fitted Treee object

Usage

# S3 method for Treee
predict(object, newdata, type = c("response", "prob", "all", "grove"), ...)

Arguments

object

a fitted model object of class Treee, which is assumed to be the result of the Treee() function.

newdata

data frame containing the values at which predictions are required. Missing values are allowed.

type

character string denoting the type of predicted value returned. The default is to return the predicted class (type = 'response'). The predicted posterior probabilities for each class will be returned if type = 'prob'. 'all' returns a data frame with predicted classes, posterior probabilities, and the predicted node indices. If cross-validation is carried out during the LDATree fitting, 'grove' option is available and will output an ensemble result from k LDATrees where k is the number of cross-validation.

...

further arguments passed to or from other methods.

Value

The function returns different values based on the type, if

type = 'response': vector of predicted responses.

type = 'prob': a data frame of the posterior probabilities. Each class takes a column.

type = 'all': a data frame contains the predicted responses, posterior probabilities, and the predicted node indices.

type = 'grove': vector of predicted responses using the ensemble method. Only available when cross-validation is carried out during the tree generating process.

Note: for factor predictors, if it contains a level which is not used to grow the tree, it will be converted to missing and will be imputed according to the missingMethod in the fitted tree.

Examples

fit <- Treee(Species~., data = iris)
#> The unpruned LDA tree is completed. For now, it has 7 nodes.
predict(fit,iris)
#>   [1] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>   [6] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [11] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [16] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [21] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [26] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [31] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [36] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [41] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [46] "setosa"     "setosa"     "setosa"     "setosa"     "setosa"    
#>  [51] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [56] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [61] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [66] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [71] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [76] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [81] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [86] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [91] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#>  [96] "versicolor" "versicolor" "versicolor" "versicolor" "versicolor"
#> [101] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [106] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [111] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [116] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [121] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [126] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [131] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [136] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [141] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
#> [146] "virginica"  "virginica"  "virginica"  "virginica"  "virginica" 
# output prosterior probabilities
predict(fit,iris,type = "prob")
#>           setosa   versicolor    virginica
#> 1   1.000000e+00 1.346467e-26 5.470122e-41
#> 2   1.000000e+00 3.044310e-22 3.780883e-36
#> 3   1.000000e+00 1.602581e-24 7.965832e-39
#> 4   1.000000e+00 2.066476e-21 1.495780e-34
#> 5   1.000000e+00 3.323135e-27 1.520898e-41
#> 6   1.000000e+00 1.957386e-24 2.347768e-38
#> 7   1.000000e+00 1.676460e-23 1.482621e-37
#> 8   1.000000e+00 1.500488e-24 2.799403e-38
#> 9   1.000000e+00 1.656470e-20 7.930489e-34
#> 10  1.000000e+00 2.464358e-23 1.515780e-36
#> 11  1.000000e+00 1.089519e-27 5.239178e-42
#> 12  1.000000e+00 4.126877e-23 3.983243e-36
#> 13  1.000000e+00 2.051133e-23 5.761935e-37
#> 14  1.000000e+00 1.153221e-25 3.454395e-40
#> 15  1.000000e+00 5.106420e-34 1.984981e-50
#> 16  1.000000e+00 4.194001e-31 1.281212e-46
#> 17  1.000000e+00 1.091899e-28 1.295337e-44
#> 18  1.000000e+00 3.081049e-25 7.068440e-40
#> 19  1.000000e+00 1.458161e-25 1.685088e-39
#> 20  1.000000e+00 1.462533e-26 7.299547e-41
#> 21  1.000000e+00 3.555888e-23 2.506082e-36
#> 22  1.000000e+00 2.090567e-24 6.680685e-39
#> 23  1.000000e+00 1.047335e-30 1.261926e-46
#> 24  1.000000e+00 9.752947e-18 2.924679e-31
#> 25  1.000000e+00 6.393547e-20 1.967620e-31
#> 26  1.000000e+00 2.643796e-20 2.584806e-33
#> 27  1.000000e+00 9.091003e-21 1.715089e-34
#> 28  1.000000e+00 1.010558e-25 1.019210e-39
#> 29  1.000000e+00 5.455611e-26 1.967406e-40
#> 30  1.000000e+00 2.482792e-21 3.934917e-34
#> 31  1.000000e+00 1.005977e-20 1.415248e-33
#> 32  1.000000e+00 1.390615e-22 3.108222e-37
#> 33  1.000000e+00 7.432336e-32 6.248063e-46
#> 34  1.000000e+00 6.420604e-33 4.068182e-48
#> 35  1.000000e+00 5.639061e-22 1.958677e-35
#> 36  1.000000e+00 3.779402e-26 2.842878e-41
#> 37  1.000000e+00 2.059638e-28 9.913396e-44
#> 38  1.000000e+00 2.238994e-28 2.317796e-42
#> 39  1.000000e+00 2.291683e-22 3.051639e-36
#> 40  1.000000e+00 9.732505e-25 1.421555e-38
#> 41  1.000000e+00 4.105188e-26 3.793648e-41
#> 42  1.000000e+00 1.262520e-15 1.790393e-29
#> 43  1.000000e+00 5.872778e-24 6.083255e-38
#> 44  1.000000e+00 7.620126e-19 4.043347e-33
#> 45  1.000000e+00 5.999323e-21 1.709602e-33
#> 46  1.000000e+00 1.073988e-20 9.621027e-35
#> 47  1.000000e+00 7.395645e-27 2.072705e-40
#> 48  1.000000e+00 2.858916e-23 5.755737e-37
#> 49  1.000000e+00 1.679742e-27 1.031727e-41
#> 50  1.000000e+00 8.100550e-25 5.403758e-39
#> 51  4.464095e-25 9.979429e-01 2.057134e-03
#> 52  1.948078e-25 9.994975e-01 5.024993e-04
#> 53  1.490783e-29 9.833311e-01 1.666888e-02
#> 54  2.969395e-26 9.998622e-01 1.377860e-04
#> 55  1.701932e-29 9.979416e-01 2.058412e-03
#> 56  3.190966e-27 9.857536e-01 1.424636e-02
#> 57  2.569378e-28 9.967943e-01 3.205701e-03
#> 58  4.596834e-15 9.999991e-01 9.091816e-07
#> 59  8.561915e-26 9.955536e-01 4.446392e-03
#> 60  6.239983e-24 9.999691e-01 3.094493e-05
#> 61  3.476106e-20 9.999882e-01 1.182807e-05
#> 62  8.883108e-25 9.999311e-01 6.891200e-05
#> 63  4.960317e-22 9.997449e-01 2.551278e-04
#> 64  3.630274e-29 9.735285e-01 2.647152e-02
#> 65  4.877106e-17 9.999995e-01 5.022658e-07
#> 66  3.026932e-23 9.998473e-01 1.527323e-04
#> 67  1.557639e-28 9.954417e-01 4.558333e-03
#> 68  1.714758e-19 9.992959e-01 7.040920e-04
#> 69  9.036601e-34 9.971284e-01 2.871597e-03
#> 70  1.082405e-20 9.999170e-01 8.298645e-05
#> 71  1.193754e-33 9.902807e-01 9.719319e-03
#> 72  3.793542e-21 9.999831e-01 1.693478e-05
#> 73  1.642963e-35 8.647550e-01 1.352450e-01
#> 74  2.850113e-27 9.118449e-01 8.815506e-02
#> 75  5.604175e-23 9.997715e-01 2.284795e-04
#> 76  3.142749e-24 9.997788e-01 2.211760e-04
#> 77  1.049613e-29 9.826809e-01 1.731905e-02
#> 78  1.633454e-34 9.692658e-01 3.073422e-02
#> 79  1.412474e-28 9.980532e-01 1.946758e-03
#> 80  4.276471e-14 9.999990e-01 1.003699e-06
#> 81  1.300528e-20 9.999621e-01 3.790260e-05
#> 82  3.443525e-18 9.999788e-01 2.116693e-05
#> 83  4.387731e-20 9.999777e-01 2.234515e-05
#> 84  0.000000e+00 1.000000e+00 0.000000e+00
#> 85  6.534357e-29 9.925844e-01 7.415583e-03
#> 86  5.879579e-26 9.994125e-01 5.875043e-04
#> 87  8.516338e-28 9.972571e-01 2.742913e-03
#> 88  5.277685e-29 9.980374e-01 1.962639e-03
#> 89  1.468551e-21 9.998580e-01 1.420047e-04
#> 90  1.158757e-24 9.998928e-01 1.071846e-04
#> 91  9.086854e-27 9.833543e-01 1.664566e-02
#> 92  2.675146e-27 9.924938e-01 7.506202e-03
#> 93  6.069960e-22 9.999197e-01 8.033337e-05
#> 94  1.134518e-15 9.999992e-01 8.070492e-07
#> 95  5.203862e-25 9.993440e-01 6.559992e-04
#> 96  4.475246e-21 9.993760e-01 6.239799e-04
#> 97  3.131526e-23 9.996004e-01 3.996085e-04
#> 98  2.357401e-23 9.996273e-01 3.727110e-04
#> 99  4.621125e-12 1.000000e+00 8.704821e-09
#> 100 5.802103e-23 9.998571e-01 1.429205e-04
#> 101 0.000000e+00 0.000000e+00 0.000000e+00
#> 102 0.000000e+00 0.000000e+00 0.000000e+00
#> 103 0.000000e+00 0.000000e+00 0.000000e+00
#> 104 0.000000e+00 0.000000e+00 0.000000e+00
#> 105 0.000000e+00 0.000000e+00 0.000000e+00
#> 106 0.000000e+00 0.000000e+00 0.000000e+00
#> 107 0.000000e+00 0.000000e+00 1.000000e+00
#> 108 0.000000e+00 0.000000e+00 0.000000e+00
#> 109 0.000000e+00 0.000000e+00 0.000000e+00
#> 110 0.000000e+00 0.000000e+00 0.000000e+00
#> 111 0.000000e+00 0.000000e+00 0.000000e+00
#> 112 0.000000e+00 0.000000e+00 0.000000e+00
#> 113 0.000000e+00 0.000000e+00 0.000000e+00
#> 114 0.000000e+00 0.000000e+00 0.000000e+00
#> 115 0.000000e+00 0.000000e+00 0.000000e+00
#> 116 0.000000e+00 0.000000e+00 0.000000e+00
#> 117 0.000000e+00 0.000000e+00 0.000000e+00
#> 118 0.000000e+00 0.000000e+00 0.000000e+00
#> 119 0.000000e+00 0.000000e+00 0.000000e+00
#> 120 0.000000e+00 0.000000e+00 1.000000e+00
#> 121 0.000000e+00 0.000000e+00 0.000000e+00
#> 122 0.000000e+00 0.000000e+00 0.000000e+00
#> 123 0.000000e+00 0.000000e+00 0.000000e+00
#> 124 0.000000e+00 0.000000e+00 0.000000e+00
#> 125 0.000000e+00 0.000000e+00 0.000000e+00
#> 126 0.000000e+00 0.000000e+00 0.000000e+00
#> 127 0.000000e+00 0.000000e+00 0.000000e+00
#> 128 0.000000e+00 0.000000e+00 0.000000e+00
#> 129 0.000000e+00 0.000000e+00 0.000000e+00
#> 130 0.000000e+00 0.000000e+00 0.000000e+00
#> 131 0.000000e+00 0.000000e+00 0.000000e+00
#> 132 0.000000e+00 0.000000e+00 0.000000e+00
#> 133 0.000000e+00 0.000000e+00 0.000000e+00
#> 134 1.663861e-35 4.810196e-01 5.189804e-01
#> 135 0.000000e+00 0.000000e+00 1.000000e+00
#> 136 0.000000e+00 0.000000e+00 0.000000e+00
#> 137 0.000000e+00 0.000000e+00 0.000000e+00
#> 138 0.000000e+00 0.000000e+00 0.000000e+00
#> 139 0.000000e+00 0.000000e+00 0.000000e+00
#> 140 0.000000e+00 0.000000e+00 0.000000e+00
#> 141 0.000000e+00 0.000000e+00 0.000000e+00
#> 142 0.000000e+00 0.000000e+00 0.000000e+00
#> 143 0.000000e+00 0.000000e+00 0.000000e+00
#> 144 0.000000e+00 0.000000e+00 0.000000e+00
#> 145 0.000000e+00 0.000000e+00 0.000000e+00
#> 146 0.000000e+00 0.000000e+00 0.000000e+00
#> 147 0.000000e+00 0.000000e+00 0.000000e+00
#> 148 0.000000e+00 0.000000e+00 0.000000e+00
#> 149 0.000000e+00 0.000000e+00 0.000000e+00
#> 150 0.000000e+00 0.000000e+00 0.000000e+00